+++ guillotina/content.py
@@ -25,6 +25,7 @@ from guillotina.events import BeforeObjectMovedEvent
 from guillotina.events import ObjectDuplicatedEvent
 from guillotina.events import ObjectLoadedEvent
 from guillotina.events import ObjectMovedEvent
+from guillotina.exceptions import ConflictError
 from guillotina.exceptions import ConflictIdOnContainer
 from guillotina.exceptions import InvalidContentType
 from guillotina.exceptions import NoPermissionToAdd
@@ -782,7 +783,7 @@ async def move(
             )
 
     if await destination_ob.async_contains(new_id):
-        raise PreconditionFailed(context, f"Destination already has object with the id {new_id}")
+        raise ConflictError(msg=f'Destination already has object with the id "{new_id}"')
 
     original_parent = context.__parent__